+++ src/main/java/net/shirojr/titanfabric/mixin/client/GameOptionsMixin.java
@@ -1,9 +1,12 @@
-package net.shirojr.titanfabric.mixin;
+package net.shirojr.titanfabric.mixin.client;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.option.GameOptions;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
+@Environment(EnvType.CLIENT)
 @Mixin(GameOptions.class)
 public abstract class GameOptionsMixin {
     @Shadow public boolean autoJump = false;
