+++ src/main/java/net/shirojr/titanfabric/mixin/client/PlayerHeldItemFeatureRendererMixin.java
@@ -1,5 +1,7 @@
-package net.shirojr.titanfabric.mixin;
+package net.shirojr.titanfabric.mixin.client;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.entity.feature.PlayerHeldItemFeatureRenderer;
 import net.minecraft.client.render.model.json.ModelTransformation;
@@ -13,6 +15,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+@Environment(EnvType.CLIENT)
 @Mixin(PlayerHeldItemFeatureRenderer.class)
 public class PlayerHeldItemFeatureRendererMixin {
     @Inject(method = "renderItem", at = @At(value = "HEAD"), cancellable = true)
